requirements.txt
pandas>=2.0
openpyxl
xlrd
pyarrow
matplotlib
seaborn
requests
python-dotenv
tqdm
